+++ src/main/java/es/uji/apps/hor/db/DepartamentoDTO.java
@@ -1,9 +1,16 @@
 package es.uji.apps.hor.db;
 
 import java.io.Serializable;
-import javax.persistence.*;
-import java.math.BigDecimal;
 import java.util.Set;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
 
 
 /**
@@ -19,7 +26,7 @@ public class DepartamentoDTO implements Serializable {
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	private Long id;
 
-	private BigDecimal activo;
+	private Long activo;
 
 	private String nombre;
 
@@ -55,11 +62,11 @@ public class DepartamentoDTO implements Serializable {
 		this.id = id;
 	}
 
-	public BigDecimal getActivo() {
+	public Long getActivo() {
 		return this.activo;
 	}
 
-	public void setActivo(BigDecimal activo) {
+	public void setActivo(Long activo) {
 		this.activo = activo;
 	}
