+++ src/main/java/es/uji/apps/hor/db/ItemCircuitoDTO.java
@@ -1,8 +1,14 @@
 package es.uji.apps.hor.db;
 
 import java.io.Serializable;
-import javax.persistence.*;
-import java.math.BigDecimal;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 
 
 /**
@@ -18,7 +24,7 @@ public class ItemCircuitoDTO implements Serializable {
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	private Long id;
 
-	private BigDecimal plazas;
+	private Long plazas;
 
 	//bi-directional many-to-one association to CircuitoDTO
     @ManyToOne
@@ -41,11 +47,11 @@ public class ItemCircuitoDTO implements Serializable {
 		this.id = id;
 	}
 
-	public BigDecimal getPlazas() {
+	public Long getPlazas() {
 		return this.plazas;
 	}
 
-	public void setPlazas(BigDecimal plazas) {
+	public void setPlazas(Long plazas) {
 		this.plazas = plazas;
 	}
